bug fix - untagging component with dependencies was showing an error saying the component is depends on itself. Also, extract the untag logic from scope.js to a new file untag-component.js
