+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- bug fix - untagging component with dependencies was showing an error saying the component is depends on itself
+
 ## [0.12.6-dev.4] - 2018-02-19
 
 - save bit dev-dependencies components inside devDependencies section of package.json
